refactor(graph): share view transforms and localize brain file labels
